feat: add ability to add custom classes to each day.
This will allow custom functionality such as differentiating what kind of
'invalid' a day might be. e.g. when used as a booking calendar, you might
not be allowed to select a given day to check in because it's sold out,
or it may simply not be a valid check-in day. Differentiating this is
key to user experience.